Split components, improve useFetchData hook
