+++ src/components/Errors/ErrorCard/TimestampKeyValue/types.ts
@@ -1,0 +1,4 @@
+export interface TimestampProps {
+  label: string;
+  timestamp: string;
+}
