+++ src/components/Errors/ErrorCard/index.tsx
@@ -1,25 +1,12 @@
-import { formatTimeDistance } from "../../../utils/formatTimeDistance";
 import { ProductionAffectionBar } from "../../Insights/InsightsCatalog/InsightsPage/insightCards/common/InsightCard/ProductionAffectionBar";
 import { CrossCircleIcon } from "../../common/icons/16px/CrossCircleIcon";
 import { Tag } from "../../common/v3/Tag";
 import { Tooltip } from "../../common/v3/Tooltip";
 import { HIGH_SEVERITY_SCORE_THRESHOLD, Score, getTagType } from "../Score";
 import { getErrorMethodId } from "../getErrorMethodId";
+import { TimestampKeyValue } from "./TimestampKeyValue";
 import * as s from "./styles";
 import { ErrorCardProps } from "./types";
-
-const renderTimestampStat = (label: string, timestamp: string) => {
-  const dateTimeString = new Date(timestamp).toString();
-
-  return (
-    <Tooltip title={dateTimeString} key={label}>
-      <s.TimestampContainer>
-        <s.TimestampLabel>{label}:</s.TimestampLabel>
-        <s.TimeDistance>{formatTimeDistance(timestamp)}</s.TimeDistance>
-      </s.TimestampContainer>
-    </Tooltip>
-  );
-};
 
 export const ErrorCard = ({ data, onClick }: ErrorCardProps) => {
   const tagType = getTagType(data.scoreInfo.score);
@@ -77,8 +64,14 @@ export const ErrorCard = ({ data, onClick }: ErrorCardProps) => {
       }
       footer={
         <s.Footer>
-          {renderTimestampStat("Started", data.firstOccurenceTime)}
-          {renderTimestampStat("Ended", data.lastOccurenceTime)}
+          <TimestampKeyValue
+            label={"Started"}
+            timestamp={data.firstOccurenceTime}
+          />
+          <TimestampKeyValue
+            label={"Ended"}
+            timestamp={data.lastOccurenceTime}
+          />
         </s.Footer>
       }
     />
