+++ src/components/Errors/ErrorDetails/ErrorDetailsCardContent/FlowStack/SpanFrameGroup/types.ts
@@ -1,0 +1,13 @@
+import { FilesURIsMap, Frame } from "../../../types";
+
+export interface FrameItemCodeLocation {
+  URI: string;
+  lineNumber: number;
+}
+
+export interface SpanFrameGroupProps {
+  spanName: string;
+  frames: Frame[];
+  filesURIs: FilesURIsMap;
+  onGoCodeLocation: (location: FrameItemCodeLocation) => void;
+}
