+++ src/components/Errors/ErrorDetails/ErrorDetailsCardContent/FlowStack/types.ts
@@ -1,12 +1,7 @@
-import { FlowInfo } from "../types";
+import { FlowInfo } from "../../types";
 
 export interface FlowProps {
   data: FlowInfo;
-}
-
-export interface FrameItemCodeLocation {
-  URI: string;
-  lineNumber: number;
 }
 
 export interface ShowOnlyWorkspaceErrorStackTraceItemsPayload {
