+++ src/components/Errors/ErrorDetails/ErrorDetailsCardContent/types.ts
@@ -1,0 +1,6 @@
+import { ErrorDetails } from "../types";
+
+export interface ErrorDetailsCardContentProps {
+  id: string;
+  data: ErrorDetails;
+}
