+++ src/components/Errors/ErrorDetails/ErrorDetailsCardHeader/types.ts
@@ -1,0 +1,6 @@
+import { ErrorDetails } from "../types";
+
+export interface ErrorDetailsCardHeaderProps {
+  onGoBack: () => void;
+  data: ErrorDetails;
+}
